Separate LVM proposal tests
